requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
packaging==23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydotplus==2.0.2
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pytz==2023.3.post1
PyYAML==6.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
uritemplate==4.1.1
zope.interface==6.0
